# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/artworks/upload": {
            "post": {
                "description": "上传图片到 CDN 并同时创建艺术作品记录",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Upload"
                ],
                "summary": "上传文件并创建作品",
                "parameters": [
                    {
                        "type": "file",
                        "description": "要上传的文件",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "作品标题",
                        "name": "title",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "艺术家名称",
                        "name": "artist",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "作品描述",
                        "name": "description",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "分类",
                        "name": "category",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "头像URL",
                        "name": "avatar_url",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/files": {
            "post": {
                "description": "上传图片或文件到 CDN",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Upload"
                ],
                "summary": "上传文件到 CDN",
                "parameters": [
                    {
                        "type": "file",
                        "description": "要上传的文件",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "上传成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.FileInfo"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "description": "从三方服务删除文件",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Upload"
                ],
                "summary": "删除文件",
                "parameters": [
                    {
                        "type": "string",
                        "description": "文件路径",
                        "name": "path",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.DeleteFileResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.DeleteFileResponse": {
            "type": "object",
            "properties": {
                "deleted": {
                    "description": "是否成功删除",
                    "type": "boolean"
                },
                "path": {
                    "description": "被删除文件路径",
                    "type": "string"
                }
            }
        },
        "models.FileInfo": {
            "type": "object",
            "properties": {
                "access_url": {
                    "description": "访问信息",
                    "type": "string"
                },
                "extra": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "file_type": {
                    "description": "文件属性",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.FileType"
                        }
                    ]
                },
                "folder_path": {
                    "description": "父目录路径",
                    "type": "string"
                },
                "metadata": {
                    "description": "元数据（用于扩展不同文件类型的信息）",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.FileMetadata"
                        }
                    ]
                },
                "modified_at": {
                    "description": "修改时间",
                    "type": "integer"
                },
                "name": {
                    "description": "基本信息",
                    "type": "string"
                },
                "original_name": {
                    "description": "原始文件名",
                    "type": "string"
                },
                "path": {
                    "description": "路径信息",
                    "type": "string"
                },
                "size": {
                    "description": "文件大小（字节）",
                    "type": "integer"
                },
                "space": {
                    "description": "空间信息",
                    "type": "string"
                },
                "storage_type": {
                    "description": "存储类型（local/smms/aliyun等）",
                    "type": "string"
                },
                "upload_time": {
                    "description": "时间信息",
                    "type": "integer"
                },
                "variants": {
                    "description": "缩略图、预览等变体",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ResourceVariant"
                    }
                }
            }
        },
        "models.FileMetadata": {
            "type": "object",
            "properties": {
                "bit_rate": {
                    "description": "音频特定",
                    "type": "integer"
                },
                "color_space": {
                    "type": "string"
                },
                "custom": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "duration": {
                    "description": "视频特定",
                    "type": "integer"
                },
                "frame_rate": {
                    "type": "number"
                },
                "hash": {
                    "type": "string"
                },
                "height": {
                    "type": "integer"
                },
                "mime_type": {
                    "description": "通用",
                    "type": "string"
                },
                "orientation": {
                    "type": "string"
                },
                "resolution": {
                    "type": "string"
                },
                "sample_rate": {
                    "type": "integer"
                },
                "video_codec": {
                    "type": "string"
                },
                "width": {
                    "description": "图片特定",
                    "type": "integer"
                }
            }
        },
        "models.FileType": {
            "type": "string",
            "enum": [
                "image",
                "video",
                "document",
                "audio",
                "other"
            ],
            "x-enum-varnames": [
                "FileTypeImage",
                "FileTypeVideo",
                "FileTypeDocument",
                "FileTypeAudio",
                "FileTypeOther"
            ]
        },
        "models.ResourceVariant": {
            "type": "object",
            "properties": {
                "access_url": {
                    "description": "访问URL",
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "extra": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "path": {
                    "description": "逻辑路径",
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                },
                "type": {
                    "description": "thumbnail, preview, compressed 等",
                    "type": "string"
                }
            }
        },
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "业务状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "响应数据"
                },
                "message": {
                    "description": "响应消息",
                    "type": "string"
                },
                "request_id": {
                    "description": "请求追踪ID",
                    "type": "string"
                },
                "timestamp": {
                    "description": "时间戳(毫秒)",
                    "type": "integer"
                }
            }
        }
    }
}